Moved updateUserProfile to updateUser
